# doc/api.yaml
openapi: 3.0.0
paths:
...
  /album/{id}:
    get:
      operationId: AlbumController_findByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Get album by id
      tags:
        - Album
    put:
      operationId: AlbumController_updateByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumCreateDto'
      responses:
        '200':
          description: Album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Update album by id
      tags:
        - Album
    delete:
      operationId: AlbumController_deleteByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success request
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete album by id
      tags:
        - Album
...
  /artist/{id}:
    get:
      operationId: ArtistController_findByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Artist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Get artist by id
      tags:
        - Artist
    put:
      operationId: ArtistController_updateByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistCreateDto'
      responses:
        '200':
          description: Artist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Update artist by id
      tags:
        - Artist
    delete:
      operationId: ArtistController_deleteByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete artist by id
      tags:
        - Artist
...
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesAddResultDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '422':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Add album to favorites by id
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete album from favorites by id
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesAddResultDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '422':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Add artist to favorites by id
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteArtist
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete artist from favorites by id
      tags:
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesAddResultDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '422':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Add track to favorites by id
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteTrack
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete track from favorites by id
      tags:
        - Favorites
  /track:
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Tracks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackDto'
      summary: Get all tracks
      tags:
        - Track
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackCreateDto'
      responses:
        '201':
          description: Track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Create track
      tags:
        - Track
  /track/{id}:
    get:
      operationId: TrackController_findByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Get track by id
      tags:
        - Track
    put:
      operationId: TrackController_updateByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackCreateDto'
      responses:
        '200':
          description: Track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Update track by id
      tags:
        - Track
    delete:
      operationId: TrackController_deleteByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      summary: Delete track by id
      tags:
        - Track
...
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:8080
components:
  schemas:
    AlbumDto:
      type: object
      properties:
        id:
          type: string
          description: UUID field
          format: uuid
        name:
          type: string
          description: Album name
        year:
          type: number
          description: Album year, should be positive
          example: 1991
          format: int32
        artistId:
          type: string
          description: Artist id
          nullable: true
          format: uuid
      required:
        - id
        - name
        - year
        - artistId
    HttpStatus:
      type: number
      enum:
        - 100
        - 101
        - 102
        - 103
        - 200
        - 201
        - 202
        - 203
        - 204
        - 205
        - 206
        - 300
        - 301
        - 302
        - 303
        - 304
        - 307
        - 308
        - 400
        - 401
        - 402
        - 403
        - 404
        - 405
        - 406
        - 407
        - 408
        - 409
        - 410
        - 411
        - 412
        - 413
        - 414
        - 415
        - 416
        - 417
        - 418
        - 421
        - 422
        - 424
        - 428
        - 429
        - 500
        - 501
        - 502
        - 503
        - 504
        - 505
    ExceptionResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Exception message
          example: User not found
        path:
          type: string
          description: Endpoint path
          example: /user/123
        timestamp:
          format: date-time
          type: string
          description: Exception timestamp
        statusCode:
          description: Exception status code
          example: 404
          allOf:
            - $ref: '#/components/schemas/HttpStatus'
      required:
        - message
        - path
        - timestamp
        - statusCode
    AlbumCreateDto:
      type: object
      properties:
        name:
          type: string
          description: Album name
        year:
          type: number
          description: Album year, should be positive
          example: 1991
          format: int32
        artistId:
          type: string
          nullable: true
          format: uuid
      required:
        - name
        - year
    ArtistDto:
      type: object
      properties:
        id:
          type: string
          description: UUID field
          format: uuid
        name:
          type: string
          description: Artist name
        grammy:
          type: boolean
          description: Is artist has grammy
      required:
        - id
        - name
        - grammy
    ArtistCreateDto:
      type: object
      properties:
        name:
          type: string
          description: Artist name
        grammy:
          type: boolean
          description: Is artist has grammy
      required:
        - name
        - grammy
    TrackDto:
      type: object
      properties:
        id:
          type: string
          description: UUID field
          format: uuid
        name:
          type: string
          description: Track name
        artistId:
          type: string
          description: Artist id
          nullable: true
          format: uuid
        albumId:
          type: string
          description: Album id
          nullable: true
          format: uuid
        duration:
          type: number
          description: Track duration
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
...
    FavoritesAddResultDto:
      type: object
      properties:
        message:
          type: string
          description: Result message
      required:
        - message
    TrackCreateDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
        duration:
          type: number
          description: Track duration
        artistId:
          type: string
          description: Artist id
          nullable: true
          format: uuid
        albumId:
          type: string
          description: Album id
          nullable: true
          format: uuid
      required:
        - name
        - duration
...